# swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Movieflix API",
        "description": "API responsavel por gerenciar os filmes da Movieflix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Buscar todos os filmes",
                "responses": {
                    "200": {
                        "description": "Retorna uma lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Cadastrar um filme",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Filme cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Já existe um filme cadastrado com esse titulo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha ao cadastrar filme"
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "summary": "Atualizar um filme",
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID do filme a ser atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Filme nao encontrado"
                    },
                    "500": {
                        "description": "Falha ao encontrar filme"
                    }
                }
            },
            "patch": {
                "summary": "Atualizar parcialmente um filme",
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID do filme a ser atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MoviePartialInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                        "description": "Filme nao encontrado"
                    },
                    "500": {
                        "description": "Falha ao atualizar filme"
                    }
                }
            },
            "delete": {
                "summary": "Deletar um filme",
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser deletado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme deletado com sucesso"
                    },
                    "404": {
                        "description": "Filme nao encontrado"
                    },
                    "500": {
                        "description": "Não foi possivel deletar o filme"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "summary": "Filtrar filmes por gênero",
                "tags": [
                    "Movie"
                ],
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "description": "Nome do gênero",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha ao filtrar filmes por gênero"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do idioma"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do idioma"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Óscar ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Óscar ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MoviePartialInput": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "Gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "Idioma do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de Óscar ganhos pelo filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            }
        }
    }
}